requirements.txt
Flask==3.0.0
gunicorn==21.2.0
alpha-vantage==3.0.0
pandas==2.2.0
numpy==1.26.3
requests==2.31.0
requests-cache==1.2.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Flask-Migrate==4.0.5
celery[redis]==5.3.4
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0
